backend:   add ability for regular user to delete site.   *disconnect if more than one user and delete if only one.
frontend:
  remove ssr check for role when displaying site controls for deletion

+++ backend/src/layers/site/site.service.ts
@@ -76,7 +76,13 @@ export class SiteService {
         return this.siteRepository.getAllFailedToday(userId);
     }
 
-    async delete(siteId: number) {
-        return this.siteRepository.delete(siteId);
+    async disconnectOrDelete(userId: number, siteId: number) {
+        const userCount = await this.siteRepository.userCount(siteId);
+
+        if (userCount === 1) {
+            return this.siteRepository.delete(siteId);
+        }
+
+        return this.siteRepository.disconnectUser(userId, siteId);
     }
 }
